requirements.txt
fastapi
uvicorn
httpx
python-dotenv
langchain
langchain-cohere
langchain-pinecone
pinecone-client
streamlit